backend/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "astrome-backend"
version = "1.0.0"
description = "Tamil Panchangam API - Swiss Ephemeris calculations for Vedic astronomy"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "AGPL-3.0"}
authors = [
    {name = "Astrome Development Team"}
]
keywords = ["panchang", "vedic", "astrology", "astronomy", "tamil", "ephemeris"]

dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyswisseph>=2.10.3.1",
    "python-dateutil>=2.8.2",
    "pytz>=2023.3",
    "timezonefinder>=6.2.0",
    "aiosqlite>=0.19.0",
    "sqlalchemy>=2.0.23",
    "alembic>=1.12.1",
    "httpx>=0.25.1",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
pythonpath = ["src"]
